-revamped blood legacies -can choose stat for blood/weapon without having them equipped -added .start command to skip that one quest everyone wants to skip (this looked like it worked, let me know if I was tripping) Will need to make a new config file and copy settings over, made quite a few changes and couldn't quite figure out how to do that automatically but it's on the list.
Not quite ready to make a release for this version since I changed a lot but should be generally good to test, try on save first and double check config before putting on live

+++ Systems/Expertise/ExpertiseSystem.cs
@@ -1,5 +1,4 @@
 ﻿using Bloodcraft.Patches;
-using Bloodcraft.Systems.Professions;
 using ProjectM;
 using ProjectM.Network;
 using Stunlock.Core;
@@ -9,9 +8,9 @@ namespace Bloodcraft.Systems.Expertise
 {
     public class ExpertiseSystem
     {
-        static readonly int UnitMultiplier = Plugin.UnitExpertiseMultiplier.Value; // Expertise points multiplier from normal units
+        static readonly float UnitMultiplier = Plugin.UnitExpertiseMultiplier.Value; // Expertise points multiplier from normal units
         static readonly int MaxExpertiseLevel = Plugin.MaxExpertiseLevel.Value; // maximum level
-        static readonly int VBloodMultiplier = Plugin.VBloodExpertiseMultiplier.Value; // Expertise points multiplier from VBlood units
+        static readonly float VBloodMultiplier = Plugin.VBloodExpertiseMultiplier.Value; // Expertise points multiplier from VBlood units
         static readonly float ExpertiseConstant = 0.1f; // constant for calculating level from xp
         static readonly int ExpertisePower = 2; // power for calculating level from xp
         public enum WeaponType
